+++ recipes/config_cache.rb
@@ -4,12 +4,6 @@
 #
 # Copyright:: 2017, The Authors, All Rights Reserved.
 
-include_recipe 'laravel-artisan::default'
-
-COMMAND = 'config:cache'
-
-execute "php artisan #{COMMAND}" do
-  cwd node['laravel-artisan']['path']
-  command "#{node['laravel-artisan']['call']} #{COMMAND} #{node['laravel-artisan']['verbosity']}"
-  action :run
+artisan 'config:cache' do
+	path node['laravel-artisan']['path']
 end