+++ recipes/schedule_run.rb
@@ -4,16 +4,8 @@
 #
 # Copyright:: 2017, The Authors, All Rights Reserved.
 
-include_recipe 'laravel-artisan::default'
-
-COMMAND = 'schedule:run'
-
-cron "laravel-scheduler" do
-	minute '*'
-    hour '*'
-    day '*'
-    month '*'
-    weekday '*'
-    command "php #{node['laravel-artisan']['path']}/artisan #{COMMAND} >> /dev/null 2>&1"
-    action :create
+### Add a crontab entry to run laravel artisan scheduler
+artisan_schedule 'schedule:run >> /dev/null 2>&1' do
+	path node['laravel-artisan']['path']
+	action :create
 end