+++ recipes/schedule_stop.rb
@@ -5,6 +5,7 @@
 # Copyright:: 2017, The Authors, All Rights Reserved.
 
 ### Remove the crontab entry for the scheduler
-cron "laravel-scheduler" do 
-    action :delete
- end
+artisan_schedule 'stop scheduler' do
+	path node['laravel-artisan']['path']
+	action :delete
+end